+++ templates/spaces/list_of_spaces.html
@@ -14,7 +14,7 @@
     <p>Look at these wonderful spaces to rent...</p>
   </div>
 
-  <a class="list-button" href="spaces/new">List a Space</a>
+  <a class="list-button" href="/spaces/new">List a Space</a>
 
   {% if spaces %}
   <div class="grid-container">
